backend/requirements.txt
faster-whisper
websockets
numpy
python-dotenv
ollama
fastapi
uvicorn
